moved the stuff that needs to be moved... kinda
